backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-multipart
python-dotenv
groq
httpx
pydantic[email]